Completed Problem 11.
Don't judge me lol. I'm sure there was a way to do this with a fraction of the code but this was easiest for me. I basically had nested loops for every set of products. a set of loops for column products, set of loops for row products, etc. I could have done it all at once if I had nested 4 loops within the column loop (j). This would allow me to get a product in 4 different directions but I would need to deal with IndexError exceptions.